requirements.txt
azure-ai-contentsafety==1.0.0
azure-ai-textanalytics==5.3.0
azure-core==1.29.2
azure-cosmos==4.5.0
azure-identity>=1.12.0
azure-keyvault-secrets>=4.6.0
fastapi==0.104.1
uvicorn==0.24.0
openai==1.59.6
python-dotenv==1.0.0
pydantic==2.10.6
tenacity==8.2.3
aiocache==0.12.2
python-multipart==0.0.9
aiohttp==3.11.14
python-dateutil==2.9.0
cachetools==5.3.3
#pydantic-core==2.32.0

